refactor(web): incorporate project home component directly into project home page
