+++ web/components/ProjectLayout/Sidebar/SidebarLink.tsx
@@ -1,5 +1,4 @@
 import Link from "next/link";
-import { useRouter } from "next/router";
 
 export interface SidebarLinkProps {
   image: React.ReactFragment;
