+++ web/pages/project/[id]/home/index.tsx
@@ -3,28 +3,100 @@ import Meta from "partials/Meta";
 import React from "react";
 import { ProjectLayout } from "hoc/ProjectLayout/ProjectLayout";
 import { useRouter } from "next/router";
-import { HomePage } from "~/components/ProjectLayout/pages/HomePage";
+import {
+  FlagIcon,
+  HornIcon,
+  PersonIcon,
+  UseableIcon,
+  GotoIcon
+} from "~/components/Icons";
 
-const Project = () => {
-    const router = useRouter();
-    const { id } = router.query;
-    const project = typeof id === "string" ? id : "";
+const WelcomeWidget: React.FC<{
+  color: string;
+  icon: typeof UseableIcon;
+  title: string;
+  description: string;
+  contentClassName?: string;
+}> = ({ color, icon, title, description, contentClassName }) => {
+  const iconHolder = { icon };
 
-    return (
-        <>
-            <Meta
-                title="Hydralite"
-                description="Hydralite is the new open source platform for project management and open source project discovery."
-                url="https://hydralite.io"
-                keywords="open source,hydralite,project management"
-            />
-            <div className={styles.project}>
-                <ProjectLayout selected={'home'} project={project}>
-                    <HomePage name={project} />
-                </ProjectLayout>
-            </div>
-        </>
-    );
+  return (
+    <div className="flex items-center justify-between h-28 w-120 bg-white-selected rounded-20 m-2 cursor-pointer select-none">
+      <div className="flex items-center">
+        <div
+          className="grid place-items-center rounded-full h-16 w-16 mx-5"
+          style={{ backgroundColor: color }}
+        >
+          {<iconHolder.icon />}
+        </div>
+        <div
+          className={`flex flex-col justify-between items-start box-border h-20 py-1 ${contentClassName}`}
+        >
+          <div className="font-bold text-2xl">{title}</div>
+          <div className="font-semibold text-base leading-5">{description}</div>
+        </div>
+      </div>
+      <div className="flex justify-center items-center w-12 h-12 mr-10 rounded-10 hover:bg-white-hover-select">
+        <div className="w-10 h-10">
+          <GotoIcon />
+        </div>
+      </div>
+    </div>
+  );
 };
 
-export default Project;
+const HomePage = () => {
+  const router = useRouter();
+  const { id } = router.query;
+  const project = typeof id === "string" ? id : "";
+
+  return (
+    <>
+      <Meta
+        title="Hydralite"
+        description="Hydralite is the new open source platform for project management and open source project discovery."
+        url="https://hydralite.io"
+        keywords="open source,hydralite,project management"
+      />
+      <div className={styles.project}>
+        <ProjectLayout selected={"home"} project={project}>
+          <div className="w-full h-full grid place-items-center">
+            <div className="">
+              <h2 className="text-4xl font-black mb-3 text-center">
+                Welcome to <span className="text-green-700">Volt</span>
+              </h2>
+              <h3 className="text-xl font-semibold text mb-7 text-center">
+                Welcome to your shiny new project. Here’s some tips to get you
+                setup.
+              </h3>
+
+              <WelcomeWidget
+                color="#FFA2D4"
+                icon={PersonIcon}
+                title="Invite your team"
+                description="Get your teammates on board, and let the collaboration begin."
+                contentClassName="w-96"
+              />
+              <WelcomeWidget
+                color="#A2B1FF"
+                icon={HornIcon}
+                title="Publicize your project"
+                description="Let others discover your project through hydra. Onboard potential contributors, all through one seamless platform."
+                contentClassName="w-112"
+              />
+              <WelcomeWidget
+                color="#FFA1A1"
+                icon={FlagIcon}
+                title="Start developing"
+                description="Go ahead and start building your project roadmap, View project traffic, and more"
+                contentClassName="w-96"
+              />
+            </div>
+          </div>
+        </ProjectLayout>
+      </div>
+    </>
+  );
+};
+
+export default HomePage;
